examples/discord-mod/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-mod"
version = "0.1.0"
description = "A DSPy project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts",
    "dspy-ai>=3.0.3",
    "dspy-cli",
    "httpx>=0.27.0",
    "py-cord[voice]>=2.6.1",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

# Use local editable install of dspy-cli for development
[tool.uv.sources]
dspy-cli = { path = "../..", editable = true }

[tool.setuptools.packages.find]
where = ["src"]
